+++ docs/conf.py
@@ -22,7 +22,7 @@ copyright = '2021, Nicolò Tampellini'
 author = 'Nicolò Tampellini'
 
 # The full version, including alpha/beta/rc tags
-release = '0.0.3'
+release = '0.0.4'
 
 
 # -- General configuration ---------------------------------------------------