+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 
 ### 0.0.2 (10 Aug 2021)
 - If pivots decrease during a bend, an exception is raised. Future versions might have a different behavior in this scenario.
+
+### 0.0.3 (10 Aug 2021)
+- setup.py bugfix.
+
+### 0.0.4 (11 Aug 2021)
+- SADDLE keyword implementation.
+- Added keywords print at top of log
+- Refactored some complex to read code